+++ src/ext_widgets/scroll_view/scroll_bar.c
@@ -477,6 +477,9 @@ static ret_t scroll_bar_get_prop(widget_t* widget, const char* name, value_t* v)
   } else if (tk_str_eq(name, WIDGET_PROP_VALUE)) {
     value_set_int(v, scroll_bar->value);
     return RET_OK;
+  } else if (tk_str_eq(name, WIDGET_PROP_AUTO_HIDE)) {
+    value_set_bool(v, scroll_bar->auto_hide);
+    return RET_OK;
   }
 
   return RET_NOT_FOUND;
@@ -498,6 +501,9 @@ static ret_t scroll_bar_set_prop(widget_t* widget, const char* name, const value
   } else if (tk_str_eq(name, WIDGET_PROP_VALUE)) {
     scroll_bar_set_value(widget, value_int(v));
     scroll_bar_update_dragger(widget);
+    return RET_OK;
+  } else if (tk_str_eq(name, WIDGET_PROP_AUTO_HIDE)) {
+    scroll_bar_set_auto_hide(widget, value_bool(v));
     return RET_OK;
   }
 
@@ -661,6 +667,8 @@ static widget_t* scroll_bar_create_internal(widget_t* parent, xy_t x, xy_t y, wh
   return_value_if_fail(scroll_bar != NULL, NULL);
 
   scroll_bar->animatable = TRUE;
+  scroll_bar->auto_hide = scroll_bar_is_mobile(widget);
+
   widget_set_state(widget, WIDGET_STATE_NORMAL);
 
   return widget;
@@ -703,6 +711,10 @@ ret_t scroll_bar_hide_by_opacity_animation(widget_t* widget, int32_t duration, i
   scroll_bar_t* scroll_bar = SCROLL_BAR(widget);
   return_value_if_fail(scroll_bar != NULL, RET_BAD_PARAMS);
 
+  if (!scroll_bar->auto_hide) {
+    return RET_OK;
+  }
+
 #ifndef WITHOUT_WIDGET_ANIMATORS
   if (scroll_bar->wa_opactiy != NULL) {
     widget_animator_destroy(scroll_bar->wa_opactiy);
@@ -735,5 +747,19 @@ ret_t scroll_bar_show_by_opacity_animation(widget_t* widget, int32_t duration, i
                      scroll_bar);
   widget_animator_opacity_set_params(scroll_bar->wa_opactiy, widget->opacity, 0xff);
   widget_animator_start(scroll_bar->wa_opactiy);
+
+  return RET_OK;
+}
+
+ret_t scroll_bar_set_auto_hide(widget_t* widget, bool_t auto_hide) {
+  scroll_bar_t* scroll_bar = SCROLL_BAR(widget);
+  return_value_if_fail(scroll_bar != NULL, RET_BAD_PARAMS);
+  assert(scroll_bar_is_mobile(widget));
+
+  scroll_bar->auto_hide = auto_hide;
+  if (!auto_hide) {
+    widget_set_visible(widget, TRUE);
+  }
+
   return RET_OK;
 }