+++ docs/changes.md
@@ -1,5 +1,7 @@
 # 最新动态
 
+2021/06/04
+  * scroll bar mobile支持auto\_hide属性。
 2021/06/03
   * 完善 scroll bar。
   * 完善 menu layout。